Added ability to decide chain splits.
The database now holds the chainWork for each block. It's the average
number of hashes that would have to be generated to create a block
chain of this difficulty.

It also now has a flag on each block and each transaction that states
whether it is part of the main chain.

By comparing the chainWork for two blocks the BlockChain object can
determine when it is necessary to switch the active chain. Note that
reorganization itself is not yet implemented.

The node now sends a corrent BlockLocator with each getblocks request.